Correct loop for comparison of host and name  It was iterating from  0 to SEEDER_COUNT. SEEDER_COUNT is a constant for the limit of the coin number. It should iterate from 0 to actualMainThreadCount. Now it iterates so. requests.

+++ db.h
@@ -20,7 +20,7 @@ extern int TempMainThreadNumber;
 
 
 extern bool bCurrentBlockFromExplorer[SEEDER_COUNT];
-extern int nCurrentBlock[SEEDER_COUNT];
+extern long long int nCurrentBlock[SEEDER_COUNT];
 extern int cfg_protocol_version[SEEDER_COUNT];
 extern int cfg_init_proto_version[SEEDER_COUNT];
 extern int cfg_min_peer_proto_version[SEEDER_COUNT];
